% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxes.R
\name{bbox_iou}
\alias{bbox_iou}
\title{The IOU of two bounding boxes}
\usage{
bbox_iou(box1, box2)
}
\arguments{
\item{box1}{numeric: 4-element box vector}

\item{box2}{matrix: nx4 matrix of boxes}
}
\value{
A vector of length n with the intersection over union (IOU) value for each box pair
}
\description{
The IOU of two bounding boxes
}
