% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{check_sha1}
\alias{check_sha1}
\title{Check file SHA1 hash against expected}
\usage{
check_sha1(filename, expected = NULL)
}
\arguments{
\item{filename}{string: file to check}

\item{expected}{string: expected SHA1 hash}
}
\value{
\code{TRUE} if the SHA1 hash matches the expected value, or if no expected value was provided.
}
\description{
Check file SHA1 hash against expected
}
