% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{ovml_class_labels}
\alias{ovml_class_labels}
\title{Class labels}
\usage{
ovml_class_labels(dataset = "coco")
}
\arguments{
\item{dataset}{string: which dataset? One of
\itemize{
\item "coco" (used with a variety of models)
\item "mvb" (used with e.g. the yolov4-mvb model)
}}
}
\value{
A character vector of class labels
}
\description{
Class labels
}
