% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pose.R
\name{process_pose_dets}
\alias{process_pose_dets}
\title{Process raw detections from pose detection network}
\usage{
process_pose_dets(
  pose,
  original_w,
  original_h,
  input_image_size,
  as = "segments",
  letterboxing = FALSE
)
}
\arguments{
\item{pose}{matrix: pose detection network output}

\item{original_w}{integer: input image width}

\item{original_h}{integer: input image height}

\item{input_image_size}{integer: network image size}

\item{as}{string: return results as "segments" or "keypoints"}

\item{letterboxing}{logical: TRUE if the input images were letterboxed to retain their original aspect ratio}
}
\value{
A data.frame
}
\description{
Process raw detections from pose detection network
}
