% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxes.R
\name{rescale_boxes}
\alias{rescale_boxes}
\title{Rescale boxes}
\usage{
rescale_boxes(
  bboxes,
  original_w,
  original_h,
  input_image_size,
  letterboxing = TRUE
)
}
\arguments{
\item{bboxes}{numeric: boxes}

\item{original_w}{numeric: original image width in pixels}

\item{original_h}{numeric: original image height in pixels}

\item{input_image_size}{numeric: the network image size in pixels}

\item{letterboxing}{logical: were the input images letterboxed?}
}
\value{
Rescaled boxes
}
\description{
Detection boxes are generally on a scaled image, with size according to the network configuration. This function takes boxes on those images and rescales back to the original image dimensions, optionally accounting for letterboxing.
}
