% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_utils.R
\name{str_trim}
\alias{str_trim}
\alias{is_num_scalar}
\alias{\%eq\%}
\alias{make_divisible}
\alias{image_wh}
\alias{image_resz}
\title{Internal utilities}
\usage{
str_trim(x)

is_num_scalar(x)

x \%eq\% y

make_divisible(x, divisor)

image_wh(im)

image_resz(im, sz, preserve_aspect = TRUE)
}
\arguments{
\item{x, y, im, sz, preserve_aspect}{: generic input parms}
}
\description{
Internal utilities
}
