% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxes.R
\name{xywh2box}
\alias{xywh2box}
\title{Convert xywh format to bounding box format}
\usage{
xywh2box(prediction)
}
\arguments{
\item{prediction}{matrix: n x m x q matrix of predictions, where the first 4 columns in q are  x y w h}
}
\value{
A matrix of the same size, with coordinates changed to xmin, ymin, xmax, ymax
}
\description{
Convert xywh format to bounding box format
}
